#862 - evaluate the node command
